updated update and delete function so only the person that made a post can change it

+++ Cooler.py
@@ -96,27 +96,31 @@ def new_note():
 @app.route('/notes/edit/<note_id>', methods=['GET', 'POST'])
 def update_note(note_id):
     # check if a user is saved in session
+    note = db.session.query(Note).filter_by(id=note_id).one()
+
     if session.get('user'):
-        if request.method == 'POST':
-            # get title data
-            title = request.form['title']
-            # get note data
-            text = request.form['noteText']
-            note = db.session.query(Note).filter_by(id=note_id).one()
-            # update note data
-            note.title = title
-            note.text = text
-            # update note int DB
-            db.session.add(note)
-            db.session.commit()
-            return redirect(url_for('get_notes'))
+        if session.get("user_id") == note.user_id:
+            if request.method == 'POST':
+                # get title data
+                title = request.form['title']
+                # get note data
+                text = request.form['noteText']
+                # update note data
+                note.title = title
+                note.text = text
+                # update note int DB
+                db.session.add(note)
+                db.session.commit()
+                return redirect(url_for('get_notes'))
 
+            else:
+                # GET request - show new note form to edit note
+
+                # retrieve note from database
+                my_note = db.session.query(Note).filter_by(id=note_id).one()
+                return render_template('new.html', note=my_note, user=session['user'])
         else:
-            # GET request - show new note form to edit note
-
-            # retrieve note from database
-            my_note = db.session.query(Note).filter_by(id=note_id).one()
-            return render_template('new.html', note=my_note, user=session['user'])
+            return redirect(url_for('get_notes'))
     else:
         # user is not in session redirect to login
         return redirect(url_for('login'))
@@ -124,14 +128,17 @@ def update_note(note_id):
 
 @app.route('/notes/delete/<note_id>', methods=['POST'])
 def delete_note(note_id):
+    my_note = db.session.query(Note).filter_by(id=note_id).one()
     # check if a user is saved in session
     if session.get('user'):
-        # retrieve note from database
-        my_note = db.session.query(Note).filter_by(id=note_id).one()
-        db.session.delete(my_note)
-        db.session.commit()
+        if session.get("user_id") == my_note.user_id:
+            # retrieve note from database
+            db.session.delete(my_note)
+            db.session.commit()
 
-        return redirect(url_for('get_notes'))
+            return redirect(url_for('get_notes'))
+        else:
+            return redirect(url_for('get_notes'))
     else:
         # user is not in session redirect to login
         return redirect(url_for('login'))
